backend/requirements.txt
asgiref
certifi
charset-normalizer
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-yasg
idna
inflection
packaging
PyJWT
python-dotenv
pytz
PyYAML
requests
setuptools
sqlparse
uritemplate
urllib3
Pillow